+++ src/Services/CRM/CRMServiceBuilder.php
@@ -18,6 +18,7 @@ use Bitrix24\SDK\Core\Credentials\Scope;
 use Bitrix24\SDK\Services\AbstractServiceBuilder;
 use Bitrix24\SDK\Services\CRM\Userfield\Service\UserfieldConstraints;
 use Bitrix24\SDK\Services\CRM\Company;
+use Bitrix24\SDK\Services\CRM\VatRates\Service\Vat;
 
 #[ApiServiceBuilderMetadata(new Scope(['crm']))]
 class CRMServiceBuilder extends AbstractServiceBuilder
@@ -26,6 +27,18 @@ class CRMServiceBuilder extends AbstractServiceBuilder
     {
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new Company\Service\CompanyDetailsConfiguration(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function vat(): VatRates\Service\Vat
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new VatRates\Service\Vat(
                 $this->core,
                 $this->log
             );