+++ src/Services/CRM/Common/Result/AbstractCrmItem.php
@@ -123,6 +123,7 @@ class AbstractCrmItem extends AbstractItem
             case 'CUSTOMIZED':
             case 'COMPLETED':
             case 'IS_INCOMING_CHANNEL':
+            case 'ACTIVE':
                 return $this->data[$offset] === 'Y';
             case 'DATE_CREATE':
             case 'CREATED_DATE':
@@ -215,6 +216,7 @@ class AbstractCrmItem extends AbstractItem
             case 'DISCOUNT_TYPE_ID':
                 return DiscountType::from($this->data[$offset]);
             case 'DISCOUNT_RATE':
+            case 'RATE':
                 return new Percentage((string)$this->data[$offset]);
             case 'TYPE_ID':
                 return ActivityType::from((int)$this->data[$offset]);