+++ CHANGELOG.md
@@ -9,6 +9,13 @@
   task «[split cli commands](https://github.com/bitrix24/b24phpsdk/issues/92)»
 - Added method `Bitrix24\SDK\Core\Credentials\Scope::getAvailableScopeCodes` returned all available scope codes, for
   task «[split cli commands](https://github.com/bitrix24/b24phpsdk/issues/92)»
+- Added service `Services\CRM\VatRates\Service\Vat` with support methods,
+  see [add crm.vat.* methods](https://github.com/bitrix24/b24phpsdk/issues/20):
+  - `get` get vat rate by id
+  - `add` add new vat rate
+  - `delete` delete vat rate
+  - `list`  get list of vat rates
+  - `update`  update vat rate
 - Developer experience: added cli command `b24-dev:show-sdk-coverage-statistics` for show actual SDK coverage for
   REST-API, see task «[split cli commands](https://github.com/bitrix24/b24phpsdk/issues/92)»
 - Developer experience: added class `Bitrix24\SDK\Deprecations\DeprecatedMethods` with list of